# docs/swagger.yml
openapi: 3.0.1
info:
  title: Vault API.
  description: "<br />
    <section>
      <h1>Introdução</h3>
      Api responsável por fazer a gestão dos PINS cadastrados para um usuário. <br />
      Todos os dados enviados e recebidos estão em formato JSON.
    </section><br /><br />
    <section>
      <h1>HTTP METHOD</h1>
      Esta API trabalha somente com os seguintes métodos HTTP.
      <ul>
        <li><b>GET: </b> Método apenas usado para buscar um recurso ou uma coleção de recursos.</li>
        <li><b>POST:</b> Método apenas usado para criação um recurso.</li>
        <li><b>PUT: </b> Método apenas usado para atualizar um recurso.</li>
      </ul>
    </section><br />
    <section>
      <h1>Internacionalização</h1>
      N/A <br />
    </section><br />
    <section>
      <h1>Permissões</h1>
      Para acessar os recursos o usuário deve ter a permissão **pin-administration** no seu escopo. Caso contrário, o acesso será barrado pelo RBAC.
    </section>
    "
  contact:
    email: ajuda@somosphi.com
  version: 1.0.0
servers:
  - url: "https://auth.api.dev.internal.phipagamentos.com/pin-vault"
    description: Ambiente de Desenvolvimento.
  - url: "https://auth.api.internal.phipagamentos.com/pin-vault"
    description: Ambiente de Produção.
tags:
  - name: RSA
    description: Collection responsável por conter os resources para obtenção da chave pública, criptografia e descriptigrafia de um PIN.
  - name: User PIN
    description: Collection responsável pela gestão dos PINS de um usuário.
paths:
  '/v1/vault/{platform}/public-key':
    get:
      summary: Busca pela chave pública
      description: Endpoint responsável por retornar a chave pública para criptografar o PIN.
      tags:
        - RSA
      parameters:
        - $ref: "#/components/parameters/platformParam"
      responses:
        "200":
          description: Retorna a chave pública em base64 e também a data de validade da chave.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSuccessBodyRSA"
              example:
                data:
                  publicKey: "PEJhc2U2NCBFbmNvZGVkIFB1YmxpYy1LZXk+"
                  expiresAt: "2011-10-05T14:48:00.000Z"
        "400":
          $ref: "#/components/responses/BAD_REQUEST"
        "500":
          $ref: "#/components/responses/INTERNAL_ERROR"
  '/v1/vault/{platform}/pins/{document}':
    get:
      summary: Valida se um determinado documento possui PIN cadastrado e ativo.
      tags:
        - User PIN
      parameters:
        - $ref: "#/components/parameters/platformParam"
        - $ref: "#/components/parameters/documentParam"
      responses:
        "200":
          description: Retorna o número do documento e a data de criação do PIN.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSuccessBodyUserPin"
              example:
                data:
                  platform: "FORTPAY"
                  document: "98765432101"
                  createdAt: "2011-10-05T14:48:00.000Z"
        "400":
          $ref: "#/components/responses/BAD_REQUEST"
        "404":
          description: Recurso não foi encontrado, podendo ser a plataforma ou o número do documento.
        "500":
          $ref: "#/components/responses/INTERNAL_ERROR"
    post:
      summary: Faz a criação de um PIN para um determinado documento.
      tags:
        - User PIN
      parameters:
        - $ref: "#/components/parameters/platformParam"
        - $ref: "#/components/parameters/documentParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptedPin"
      responses:
        "204":
          description: Registro criado com sucesso.
        "400":
          $ref: "#/components/responses/BAD_REQUEST"
        "500":
          $ref: "#/components/responses/INTERNAL_ERROR"

  '/v1/vault/{platform}/pins/{document}/verify':
    post:
      summary: Verifica se um PIN está cadastrado e ativo para um determinado documento.
      tags:
        - User PIN
      parameters:
        - $ref: "#/components/parameters/platformParam"
        - $ref: "#/components/parameters/documentParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptedPin"
      responses:
        "204":
          description: Registro criado com sucesso.
        "400":
          $ref: "#/components/responses/BAD_REQUEST"
        "401":
          $ref: "#/components/responses/UNAUTHORIZED"
        "500":
          $ref: "#/components/responses/INTERNAL_ERROR"
components:
  parameters:
    platformParam:
      name: platform
      in: path
      required: true
      schema:
        type: string
      description: Nome da plataforma a ser consultada.
    documentParam:
      name: document
      in: path
      required: true
      schema:
        type: string
      description: Número do documento.
  schemas:
    ResponseSuccessBodyRSA:
      description: Conteúdo que representa o recurso solicitado pelo Client.
      type: object
      properties:
        data:
          type: object
          properties:
            publicKey:
              description: Chave pública codificada em base64
              type: string
              format: byte
            expiresAt:
              description: Data de validade da chave pública. Formato ISO-8601
              type: string
              format: date-time
    ResponseSuccessBodyUserPin:
      description: Conteúdo que representa o recurso solicitado pelo Client.
      type: object
      properties:
        data:
          type: object
          properties:
            platform:
              description: Nome da plataforma
              type: string
            document:
              description: Número do documento
              type: string
            createdAt:
              description: Data de criação do PIN. Formato ISO-8601
              type: string
              format: date-time
    ResponseErrorBody:
      description: Objeto que representa o resultado de toda requisição que possui algum erro.
      type: object
      properties:
        errors:
          description: Lista com os erros ocorridos.
          type: array
          items:
            properties:
              code:
                description: Código do erro.
                type: string
              title:
                description: Descrição do status de erro.
                type: string
              detail:
                description: Mensagem descritiva sobre o erro.
                type: string
              fields:
                description: Campos relacionados ao erro ocorrido.
                type: array
                items:
                  type: string
    EncryptedPin:
      required:
        - encryptedPin
      type: object
      properties:
        encryptedPin:
          type: string
          format: byte
          description: Código base64 do PIN encriptado com a chave pública.
      example:
        encryptedPin: "PEJhc2U2NCBFbmNvZGVkIFB1YmxpYy1LZXk+"
  responses:
    UNAUTHORIZED:
      description: Você não tem permissão(scope) necessária para poder visualizar este resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseErrorBody"
          example:
            errors: [
              {
                code: "401",
                title: "Unauthorized",
                detail: "<GENERIC ERROR MESSAGE>",
                fields: []
              }
            ]
    BAD_REQUEST:
      description: Requisição não válida para a aplicação.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseErrorBody"
          example:
            errors: [
              {
                code: "400",
                title: "Bad request",
                detail: "<GENERIC ERROR MESSAGE>",
                fields: [
                    "fieldName1",
                    "fieldName2",
                    "fieldName3"
                ]
              }
            ]
    INTERNAL_ERROR:
      description: Aconteceu um erro inesperado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseErrorBody"
          example:
            errors: [
              {
                code: "500",
                title: "Internal server error",
                detail: "<GENERIC ERROR MESSAGE>",
                fields: [
                    "fieldName1"
                ]
              }
            ]